Win32: Fix MSYS2/GCC compilation errors
It does not figure out what a size_t is without specific headers, but it did know you're supposed to include cstddef (which I'd rather avoid).

+++ Core/RagLite2.hpp
@@ -44,10 +44,10 @@
 #define EXPAND_AS_STRING(x) #x
 #define TOSTRING(x) EXPAND_AS_STRING(x)
 
-constexpr size_t BITS_PER_BYTE = 8ULL;
+constexpr int BITS_PER_BYTE = 8;
 #define Bits(bits) ((bits) / BITS_PER_BYTE)
 
-constexpr size_t PLATFORM_POINTER_SIZE = sizeof(void*);
+constexpr int PLATFORM_POINTER_SIZE = sizeof(void*);
 static_assert(PLATFORM_POINTER_SIZE == Bits(64), "Only 64-bit platforms are currently supported");
 
 #if defined(__BYTE_ORDER__) && __BYTE_ORDER__ == __ORDER_BIG_ENDIAN__
